fix: correct pgl::int128 Boost fallback semantics for MSVC
The Boost.Multiprecision fallback for pgl::int128 (used when native __int128
is unavailable, e.g. MSVC) did not faithfully reproduce __int128, which broke
the geometry stack in two ways uncovered by the new Windows CI job:

- Range mismatch: Boost's int128_t is sign-magnitude with a 128-bit magnitude,
  so its max() is 2^128-1 rather than __int128's 2^127-1. pgl::int128 silently
  had a different range on the fallback, breaking shared overflow guards and
  the BigInt string tests. Switch to a 127-bit magnitude backend whose max()
  matches __int128 exactly (min() differs only at -2^127, never relied upon).

- No double interop: Boost's integer backend omits mixed operations with
  floating-point types, which __int128 supports via the usual arithmetic
  conversions. Add a global-scope shim (comparison + arithmetic) so pgl::int128
  stays a drop-in. A single operator<=> and operator== avoid C++20
  reversed-candidate ambiguity.

The missing double interop also caused a real infinite recursion (observed as
a stack overflow on MSVC): Rational(Float)'s `requires(Int x, Float g){x*g;}`
branch was false for the Boost int128, so it took the else branch that
delegates to Rational<int128>(f) -- which recursed into itself forever. With
the shim the branch is taken directly (base case); rational.hpp wraps the
product in Int() so the resulting double assigns back to the Boost int128.

Also rename local `far` variables in 6 tests (`far`, like `near`, is a
reserved MSVC macro).

Verified: native g++/clang 46/46; the full suite under a forced Boost-int128
build (g++ -U__SIZEOF_INT128__) 45/46, the one gap being a vendored doctest
2.5.0 / Boost stringification quirk under g++ only.

+++ include/core/rational.hpp
@@ -180,7 +180,11 @@ public:
                 digits -= exponent;
             assert(digits > 0); // Rational number overflow
             den = Int(1) << digits;
-            num = negative ? -den * f : den * f;
+            // Wrap in Int(): for the Boost int128 fallback `den * f` yields a
+            // double (see the double-interop shim in numeric.hpp), so convert it
+            // back explicitly. For native __int128 / built-in ints this is the
+            // same truncating conversion that the implicit assignment did.
+            num = Int(negative ? -den * f : den * f);
 
             normalized_ = false;
             normalize();
